fix PR review #JC19-354
