+++ apps/vanilla/src/pages/profile/profile.ts
@@ -2,14 +2,14 @@ import { User } from '@js-camp/core/models/user/user';
 import { assertNonNull } from '@js-camp/core/utils/assertNonNull';
 
 import { Navbar } from '../../namespaces/navbar';
-import { Utility } from '../../namespaces/utility';
 
 import { AuthService } from '../../services/authService';
 import { UserService } from '../../services/userService';
+import { convertDate } from '../../utils/convertDate';
 
 window.addEventListener('load', async(): Promise<void> => {
   await AuthService.navigateByAuthorization();
-  await Navbar.renderNavbar();
+  await Navbar.render();
   renderUserProfile();
 });
 
@@ -31,8 +31,8 @@ export function renderDetail(user: User): void {
     <li>Email: ${user.email}</li>
     <li>First Name: ${user.firstName ?? 'No first name available'}</li>
     <li>Last Name: ${user.lastName ? user.lastName : 'No last name available'}</li>
-    <li>Created at: ${Utility.convertDate(user.created)}</li>
-    <li>Modified at: ${Utility.convertDate(user.modified)}</li>
+    <li>Created at: ${convertDate(user.created)}</li>
+    <li>Modified at: ${convertDate(user.modified)}</li>
     <li>Avatar: ${user.avatar ? `
       <img class="user-avatar" src="${user.avatar}" alt="${user.firstName}${user.lastName}" />
     ` : 'No avatar available'}
