+++ apps/vanilla/src/utils/hasSortOption.ts
@@ -1,0 +1,7 @@
+/**
+ * Checks whether has sort options enabled.
+ * @param option Option value to check.
+ */
+export function hasSortOption(option: string): boolean {
+  return option !== '';
+}
