+++ apps/vanilla/src/utils/setDirectionState.ts
@@ -1,0 +1,17 @@
+import { assertNonNull } from '@js-camp/core/utils/assertNonNull';
+
+/**
+ * Change state of the order direction.
+ * @param hasOption Whether has sort option enabled.
+ */
+export function setDirectionState(hasOption: boolean): void {
+  const sortDirection = document.querySelector<HTMLSelectElement>('.sort__direction');
+  assertNonNull(sortDirection);
+  if (hasOption) {
+    sortDirection.disabled = false;
+    sortDirection.style.cursor = 'default';
+  } else {
+    sortDirection.disabled = true;
+    sortDirection.style.cursor = 'not-allowed';
+  }
+}
